requirements.txt
langchain_core==0.3.0
langchain_openai==0.2.0
pypdf==4.3.1
pytest==8.2.2
pyperclip==1.8.2
python-dotenv==1.0.1
requests==2.32.3
streamlit
